requirements.txt
Flask
Flask-Cors
Flask-SocketIO
Flask-SQLAlchemy
eventlet
gunicorn
python-dotenv
psycopg2-binary
gspread
pytest
pytest-cov
cloudinary